Automate_Webscraping/requirements.txt
pandas==2.2.2
selenium==4.23.0
beautifulsoup4==4.12.3
lxml==5.2.2
html5lib==1.1
python-dotenv==1.0.1
SQLAlchemy==1.4.52
mysql-connector-python==8.4.0
